Refactor: Layout 리팩토링
